+++ profiler2/__main__.py
@@ -42,18 +42,6 @@ import sys
 
 from . import helpers, manager
 
-# hard set no support for non linux platforms
-if "linux" not in sys.platform:
-    sys.exit("{0} only works on Linux... exiting...".format(os.path.basename(__file__)))
-
-# hard set no support for python < v3.7
-if sys.version_info < (3, 7):
-    sys.exit(
-        "{0} requires Python version 3.7 or higher...\nyou are trying to run with Python version {1}...\nexiting...".format(
-            os.path.basename(__file__), platform.python_version()
-        )
-    )
-
 
 def main() -> None:
     """ Set up args and start the profiler manager """
@@ -62,5 +50,23 @@ def main() -> None:
     manager.start(args)
 
 
-if __name__ == "__main__":
-    main()
+def init():
+    # hard set no support for non linux platforms
+    if "linux" not in sys.platform:
+        sys.exit(
+            "{0} only works on Linux... exiting...".format(os.path.basename(__file__))
+        )
+
+    # hard set no support for python < v3.7
+    if sys.version_info < (3, 7):
+        sys.exit(
+            "{0} requires Python version 3.7 or higher...\nyou are trying to run with Python version {1}...\nexiting...".format(
+                os.path.basename(__file__), platform.python_version()
+            )
+        )
+
+    if __name__ == "__main__":
+        sys.exit(main())
+
+
+init()